requirements.txt
# Web Framework and Dependencies
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2

# Machine Learning and Image Processing
tensorflow==2.13.0
tensorflow-intel==2.13.0
keras==2.13.1
opencv-python==4.8.0.76
Pillow==10.0.0

# Data Processing and Visualization
numpy==1.24.3
matplotlib==3.7.2
scipy==1.11.2

# PDF Generation
xhtml2pdf==0.2.11
reportlab==4.0.7

# TensorFlow Dependencies
h5py==3.9.0
absl-py==2.0.0
grpcio==1.56.2
tensorboard==2.13.0
protobuf==4.24.3
typing-extensions==4.5.0
